refactor(validation): split the line-ending record out of the gate
Pure move, no behavior change. `check_index_line_endings.py` reached 485 of
the 500-line `file-size` ceiling and `taste-lints` was warning at that
distance, so the next fix could not land without extraction.

The seam is where the subject changes. `index_line_endings_record.py` holds
what a `git ls-files --eol` row is and how a tracked pathname renders:
`_BAD_INDEX_STATES`, `_LF_ATTRIBUTES`, `Violation`, `parse_violations`,
`display_path`, `is_spellable`. None of it runs a subprocess or touches a
repository. The gate keeps everything that invokes git, decides, reports and
remediates, and re-exports the three names it still calls.

`display_path` moved with `Violation.render`, its only caller in the gate, so
the two unit tests over it in `test_check_index_line_endings_fix.py` now name
the module that owns it.

Refs #5475
